requirements.txt
alembic==1.11.3
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.2
flask-apidoc==1.3.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-dependency==0.5.1
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
WTForms==3.0.1
